Revert "Feat: Stripe FE"
