+++ Clients/src/application/repository/tiers.repository.ts
@@ -1,13 +1,5 @@
 import { apiServices } from "../../infrastructure/api/networkServices";
 import { getAuthToken } from "../redux/auth/getAuthToken";
-import { Tier } from "../../domain/types/Tiers";
-
-interface TierResponse {
-  id: number;
-  name: string;
-  price: number;
-  features: Tier;
-}
 
 /**
  * Retrieves features for a specific tier.
@@ -36,19 +28,4 @@ export async function getTierFeatures({
     responseType,
   });
   return response.data;
-}
-
-export async function getAllTiers({
-  signal,
-  authToken = getAuthToken(),
-}: {
-  signal?: AbortSignal;
-  authToken?: string;
-} = {}): Promise<TierResponse[]> {
-  const response = await apiServices.get("/tiers", {
-    headers: { Authorization: `Bearer ${authToken}` },
-    signal,
-  });
-
-  return response.data.data ?? [];
 }
