+++ Servers/domain.layer/models/organization/organization.model.ts
@@ -159,7 +159,6 @@ export class OrganizationModel
   async updateOrganization(updateData: {
     name?: string;
     logo?: string;
-    subscription_id?: number;
     members?: number[];
     projects?: number[];
   }): Promise<void> {
@@ -216,10 +215,6 @@ export class OrganizationModel
       }
 
       this.logo = updateData.logo || "";
-    }
-
-    if (updateData.subscription_id !== undefined) {
-      this.subscription_id = updateData.subscription_id;
     }
   }
